+++ CHANGELOG.md
@@ -7,6 +7,8 @@
   - Add chat channels for frozen players
   - Add ipbans-migrate command (migrate default bukkit bans to staff++)
   - Add bans-migrate command (migrate default bukkit ipbans to staff++)
+  - Add configured commands on chat phrase detection
+  - The report command can now be used from console.
 
 ### Bugs
   - Fix potion effects not wearing off if a frozen player logs out.